requirements.txt
# local package
-e .

# requirements
pillow == 10.2.0
bestconfig == 1.3.6
tqdm == 4.66.2
numpy == 1.26.3
matplotlib~=3.8.4
pandas == 2.2.1
scipy == 1.13.0
scikit-image == 0.23.1
torch == 2.2.2
torchvision == 0.17.2
ultralytics == 8.1.35
transformers == 4.39.1
segmentation-models-pytorch == 0.3.3
timm == 0.9.2
segment-anything == 1.0
opencv-python == 4.9.0.80
onnx == 1.16.0
onnxruntime == 1.16.0
onnxruntime-gpu == 1.16.0
albumentations == 1.4.2
pytest == 8.1.1
mlflow == 2.11.3

# external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1

click~=8.1.7
PyYAML~=6.0.1
setuptools~=68.2.0